Remove `which` as a dependency
In Ruby 3, if which is not present on the system, `which timeout` will
raise an Errno::ENOENT. This can be suppressed by appending
2>/dev/null to the command. The which command is not necessary as we
are using the shell to find it anyway, what matters is the existence
or otherwise of the timeout command. Errno::ENOENT is still caught
just in case a future version of Ruby decides to change its mind.